Refactor(dto): 기본 정보 dto와 결과 dto를 분리
