+++ src/main/java/dto/PlayerInfoDto.java
@@ -1,20 +1,14 @@
 package dto;
 
-import domain.participant.WinStatus;
-
 import java.util.List;
 
 public class PlayerInfoDto {
-    private String name;
-    private List<String> cards;
-    private int score;
-    private WinStatus winStatus;
+    private final String name;
+    private final List<String> cards;
 
-    public PlayerInfoDto(String name, List<String> cards, int score, WinStatus winStatus) {
+    public PlayerInfoDto(String name, List<String> cards) {
         this.name = name;
         this.cards = cards;
-        this.score = score;
-        this.winStatus = winStatus;
     }
 
     public String getName() {
@@ -23,13 +17,5 @@ public class PlayerInfoDto {
 
     public List<String> getCards() {
         return cards;
-    }
-
-    public int getScore() {
-        return score;
-    }
-
-    public WinStatus getWinStatus() {
-        return winStatus;
     }
 }
